+++ src/containers/SettingsGeneral.tsx
@@ -11,6 +11,7 @@ import {
   toggleAutoFocusInput,
   toggleCopyMessageOnDoubleClick,
   toggleDisableDialogAnimations,
+  toggleEnablePollEditor,
   toggleHideWhispers,
   toggleHighlightAllMentions,
   togglePrioritizeUsernames,
@@ -23,6 +24,7 @@ import {
   getAutoFocusInput,
   getCopyMessageOnDoubleClick,
   getDisableDialogAnimations,
+  getEnablePollEditor,
   getHideWhispers,
   getHighlightAllMentions,
   getPrioritizeUsernames,
@@ -68,6 +70,7 @@ class SettingsGeneral extends React.Component<Props> {
       autoFocusInput,
       copyMessageOnDoubleClick,
       disableDialogAnimations,
+      enablePollEditor,
       hideWhispers,
       highlightAllMentions,
       prioritizeUsernames,
@@ -127,6 +130,12 @@ class SettingsGeneral extends React.Component<Props> {
           />
         </SettingsPanelSection>
         <SettingsPanelSection title="Streamer mode">
+          <Switch
+            description="This feature uses Straw Poll."
+            checked={enablePollEditor}
+            label="Enable poll editor"
+            onChange={this.props.toggleEnablePollEditor}
+          />
           <Switch
             description="You will still be able to send whispers."
             checked={hideWhispers}
@@ -198,6 +207,7 @@ export default connect<StateProps, DispatchProps, {}, ApplicationState>(
     autoFocusInput: getAutoFocusInput(state),
     copyMessageOnDoubleClick: getCopyMessageOnDoubleClick(state),
     disableDialogAnimations: getDisableDialogAnimations(state),
+    enablePollEditor: getEnablePollEditor(state),
     hideWhispers: getHideWhispers(state),
     highlightAllMentions: getHighlightAllMentions(state),
     prioritizeUsernames: getPrioritizeUsernames(state),
@@ -209,6 +219,7 @@ export default connect<StateProps, DispatchProps, {}, ApplicationState>(
     toggleAutoFocusInput,
     toggleCopyMessageOnDoubleClick,
     toggleDisableDialogAnimations,
+    toggleEnablePollEditor,
     toggleHideWhispers,
     toggleHighlightAllMentions,
     togglePrioritizeUsernames,
@@ -225,6 +236,7 @@ type StateProps = {
   autoFocusInput: ReturnType<typeof getAutoFocusInput>
   copyMessageOnDoubleClick: ReturnType<typeof getCopyMessageOnDoubleClick>
   disableDialogAnimations: ReturnType<typeof getDisableDialogAnimations>
+  enablePollEditor: ReturnType<typeof getEnablePollEditor>
   hideWhispers: ReturnType<typeof getHideWhispers>
   highlightAllMentions: ReturnType<typeof getHighlightAllMentions>
   prioritizeUsernames: ReturnType<typeof getPrioritizeUsernames>
@@ -240,6 +252,7 @@ type DispatchProps = {
   toggleAutoFocusInput: typeof toggleAutoFocusInput
   toggleCopyMessageOnDoubleClick: typeof toggleCopyMessageOnDoubleClick
   toggleDisableDialogAnimations: typeof toggleDisableDialogAnimations
+  toggleEnablePollEditor: typeof toggleEnablePollEditor
   toggleHideWhispers: typeof toggleHideWhispers
   toggleHighlightAllMentions: typeof toggleHighlightAllMentions
   togglePrioritizeUsernames: typeof togglePrioritizeUsernames